requirements.txt
# Dev Knowledge Graph dependencies
rdflib>=7.0.0
requests>=2.31.0
google-cloud-aiplatform>=1.74.0
google-auth>=2.27.0
python-dotenv>=1.0.0
qwikidata>=0.3.4
google-cloud-storage>=2.14.0
watchdog>=4.0.0
langchain-google-genai>=2.0.0
langgraph>=0.2.0
pydantic>=2.0.0
